+++ filepairs/content_pair_matcher_test.go
@@ -1,3 +1,4 @@
+//nolint:dupl
 package filepairs_test
 
 import (