+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - gitseek: Add `StatusLangFileUpToDate` constant for up-to-date language files
 - filepairs: Generate pairs from EN files and include missing language files
 - web: Exclude `lang-file-missing` and `up-to-date` from default dashboard filters
+- filepairs: add i18n pair pattern for i18n/<lang>/<lang>.toml
 
 ## [v0.0.0] - 2026-03-14